+++ src/Oro/Bundle/SearchBundle/Tests/Unit/DependencyInjection/OroSearchExtensionTest.php
@@ -5,25 +5,10 @@ namespace Oro\Bundle\SearchBundle\Tests\Unit\DependencyInjection;
 use Oro\Bundle\SearchBundle\Controller\Api\SearchAdvancedController;
 use Oro\Bundle\SearchBundle\Controller\Api\SearchController;
 use Oro\Bundle\SearchBundle\DependencyInjection\OroSearchExtension;
-use Oro\Bundle\SearchBundle\Tests\Unit\Fixture\Bundle\FirstEngineBundle\FirstEngineBundle;
-use Oro\Bundle\SearchBundle\Tests\Unit\Fixture\Bundle\SecondEngineBundle\SecondEngineBundle;
-use Oro\Component\Config\CumulativeResourceManager;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 class OroSearchExtensionTest extends \PHPUnit\Framework\TestCase
 {
-    protected function setUp(): void
-    {
-        $bundle1 = new FirstEngineBundle();
-        $bundle2 = new SecondEngineBundle();
-        CumulativeResourceManager::getInstance()
-            ->clear()
-            ->setBundles([
-                $bundle1->getName() => get_class($bundle1),
-                $bundle2->getName() => get_class($bundle2)
-            ]);
-    }
-
     public function testLoadDefaultConfig(): void
     {
         $container = new ContainerBuilder();
@@ -32,8 +17,8 @@ class OroSearchExtensionTest extends \PHPUnit\Framework\TestCase
         $extension->load([], $container);
 
         self::assertEquals(
-            'orm',
-            $container->getParameter('oro_search.engine')
+            'orm:',
+            $container->getParameter('oro_search.engine_dsn')
         );
         self::assertSame(
             [],
@@ -56,7 +41,7 @@ class OroSearchExtensionTest extends \PHPUnit\Framework\TestCase
         $container = new ContainerBuilder();
 
         $config = [
-            'engine'            => 'some-other-engine',
+            'engine_dsn'        => 'some-other-engine:',
             'engine_parameters' => ['some-engine-parameters'],
             'log_queries'       => true
         ];
@@ -65,8 +50,8 @@ class OroSearchExtensionTest extends \PHPUnit\Framework\TestCase
         $extension->load(['oro_search' => $config], $container);
 
         self::assertEquals(
-            $config['engine'],
-            $container->getParameter('oro_search.engine')
+            $config['engine_dsn'],
+            $container->getParameter('oro_search.engine_dsn')
         );
         self::assertEquals(
             $config['engine_parameters'],
@@ -80,35 +65,5 @@ class OroSearchExtensionTest extends \PHPUnit\Framework\TestCase
             '@OroSearch/Datagrid/itemContainer.html.twig',
             $container->getParameter('oro_search.twig.item_container_template')
         );
-    }
-
-    public function testOrmSearchEngineLoad(): void
-    {
-        $container = new ContainerBuilder();
-
-        $config = [
-            'engine' => 'orm'
-        ];
-
-        $extension = new OroSearchExtension();
-        $extension->load(['oro_search' => $config], $container);
-
-        self::assertTrue($container->hasDefinition('test_orm_service'));
-    }
-
-    public function testOtherSearchEngineLoad(): void
-    {
-        $container = new ContainerBuilder();
-
-        $config = [
-            'engine' => 'other_engine'
-        ];
-
-        $extension = new OroSearchExtension();
-        $extension->load(['oro_search' => $config], $container);
-
-        self::assertTrue($container->hasDefinition('test_engine_service'));
-        self::assertTrue($container->hasDefinition('test_engine_first_bundle_service'));
-        self::assertTrue($container->hasDefinition('test_engine_second_bundle_service'));
     }
 }