+++ src/Oro/Bundle/SearchBundle/Tests/Unit/Engine/ObjectMapperTest.php
@@ -209,6 +209,7 @@ class ObjectMapperTest extends \PHPUnit\Framework\TestCase
             $configProvider,
             $cache,
             'test',
+            'test',
             'test'
         );